refactor: split terminal launch and login shell modules
keeps the max-lines ratchet satisfied by separating concerns instead of
refreshing the baseline: login shell launch conversion moves to its own
dart part and test group, and the default launch construction moves out
of terminal_spawn.rs into terminal_launch_defaults.rs.
