+++ lib/src/features/workbench/presentation/terminal_runtime.dart
@@ -34,6 +34,7 @@ part 'terminal_runtime_output_batching.dart';
 part 'terminal_runtime_startup_delivery.dart';
 part 'terminal_runtime_interactive_view.dart';
 part 'terminal_runtime_shell_launches.dart';
+part 'terminal_login_shell_launch.dart';
 part 'terminal_runtime_posix_io.dart';
 part 'terminal_runtime_testing.dart';
